backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
pydantic==2.11.1
requests==2.32.3
praw==7.8.1
litellm==1.67.0
psutil==5.9.8
